# cmd/docs/swagger.yaml
definitions:
  dto.UpdateCategoryRequestDTO:
    properties:
      categoryName:
        type: string
    type: object
  github_com_sankangkin_di-rest-api_internal_models.Category:
    properties:
      categoryName:
        minLength: 3
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      updatedAt:
        type: string
    required:
    - categoryName
    type: object
...
  github_com_sankangkin_di-rest-api_internal_models.Role:
    enum:
    - admin
    - user
    type: string
    x-enum-varnames:
    - ADMIN
    - USER
...
  github_com_sankangkin_di-rest-api_internal_models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      isAdmin:
        type: boolean
      password:
        minLength: 3
        type: string
      role:
        $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Role'
      updatedAt:
        type: string
      userName:
        minLength: 3
        type: string
    required:
    - email
    - isAdmin
    - password
    - role
    - userName
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  httputil.HttpError400:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  httputil.HttpError401:
    properties:
      code:
        example: 401
        type: integer
      message:
        example: Unauthorized
        type: string
    type: object
  httputil.HttpError500:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: internal server error
        type: string
    type: object
  internal_auth.RefreshRequestDTO:
    properties:
      refreshToken:
        type: string
    type: object
  internal_auth.RefreshResponseDTO:
    properties:
      refreshToken:
        type: string
    type: object
  internal_auth.SignInRequestDTO:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_auth.SignInResponseDTO:
    properties:
      accessToken:
        type: string
      refreshToken:
        type: string
      role:
        type: string
      userName:
        type: string
    type: object
  internal_auth.SignUpDTO:
    properties:
      email:
        type: string
      password:
        type: string
      role:
        type: string
      userName:
        type: string
    required:
    - email
    - password
    - role
    - userName
    type: object
  internal_auth.SignUpResponseDTO:
    properties:
      AccessToken:
        type: string
      RefreshToken:
        type: string
      user:
        $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.User'
    type: object
  internal_domain_category.CreateCategoryRequestDTO:
    properties:
      categoryName:
        type: string
    type: object
...
host: localhost:5555
info:
  contact:
    email: sankaungkin@gmail.com
    name: San KaungKin
  description: This is an auto-generated API docs.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: REST-API with(golang fiber, google wire dependency injection)
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Data
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/internal_auth.SignInRequestDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.SignInResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      summary: Login to the api with email and password
      tags:
      - Auth
  /api/auth/logout:
    post:
      description: Logout user
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      summary: Logout user
      tags:
      - Auth
  /api/auth/refresh:
    post:
      consumes:
      - application/json
      description: Get new access token using refresh token
      parameters:
      - description: Refresh token request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.RefreshResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
      summary: Refresh access token
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register new user based on parameters
      parameters:
      - description: Signup Data
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/internal_auth.SignUpDTO'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.SignUpResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      summary: Create new user based on parameters
      tags:
      - Auth
  /api/categories:
    get:
      consumes:
      - application/json
      description: Fetch all Categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Category'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - ApiKeyAuth: []
      - Bearer  <-----------------------------------------add this in all controllers that need authentication: []
      summary: Fetch all Categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create new category based on parameters
      parameters:
      - description: Category Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/internal_domain_category.CreateCategoryRequestDTO'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - ApiKeyAuth: []
      - Bearer  <-----------------------------------------add this in all controllers that need authentication: []
      summary: Create new category based on parameters
      tags:
      - Categories
  /api/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete individual category
      parameters:
      - description: category Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - ApiKeyAuth: []
      - Bearer  <-----------------------------------------add this in all controllers that need authentication: []
      summary: Delete individual category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Fetch individual category by Id
      parameters:
      - description: category Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - ApiKeyAuth: []
      - Bearer  <-----------------------------------------add this in all controllers that need authentication: []
      summary: Fetch individual category by Id
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update individual category
      parameters:
      - description: category Id
        in: path
        name: id
        required: true
        type: string
      - description: Category Data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCategoryRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_sankangkin_di-rest-api_internal_models.Category'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HttpError400'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HttpError401'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HttpError500'
      security:
      - ApiKeyAuth: []
      - Bearer  <-----------------------------------------add this in all controllers that need authentication: []
      summary: Update individual category
      tags:
      - Categories
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"